requirements.txt
alembic==1.10.2
aniso8601==9.0.1
APScheduler==3.10.1
apturl==0.5.2
attrs==22.2.0
aws-cdk-lib==2.90.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
aws-ssm==0.4.0
bcrypt==4.0.1
blinker==1.6
boto3==1.26.104
botocore==1.29.104
Brlapi==0.8.3
cattrs==23.1.2
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
command-not-found==0.3
constructs==10.2.69
cryptography==40.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
docopt==0.6.2
docusign-esign==3.21.0
ecdsa==0.18.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
greenlet==2.0.2
gunicorn==20.1.0
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
importlib-resources==6.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsii==1.86.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==8.10.0
mypy==1.5.0
mypy-extensions==1.0.0
netifaces==0.11.0
nose==1.3.7
oauthlib==3.2.0
olefile==0.46
packaging==23.0
paramiko==3.2.0
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.12.4
ptyprocess==0.7.0
publication==0.0.3
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
pyjarowinkler==1.8
PyJWT==2.6.0
pymacaroons==0.13.0
PyMySQL==1.0.3
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==3.10.0
PyPDF2==3.0.1
pyRFC3339==1.1
pysftp==0.2.9
pytest==7.2.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-jose==3.3.0
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
SecretStorage==3.3.1
sentry-sdk==1.18.0
six==1.16.0
SQLAlchemy==1.4.0b1
systemd-python==234
typeguard==2.13.3
typing_extensions==4.5.0
tzlocal==5.0.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.15
wadllib==1.3.6
Werkzeug==2.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
